Component Renames. Splash Image Centering.
