+++ navigators/AppNavigator.js
@@ -1,6 +1,6 @@
 import { TabNavigator } from 'react-navigation';
 
-import HomeScreen from '../components/Home.js';
+import HomeScreen from '../components/HomeScreen.js';
 
 const AppNavigator = TabNavigator({
     Home: { screen: HomeScreen},
