+++ reducers/NavigationReducer.js
@@ -1,11 +1,11 @@
-import { RootNavigator } from '../navigators/root.js';
+import { RootNavigator } from '../navigators/RootNavigator.js';
 
 const splashAction = RootNavigator.router.getActionForPathAndParams('Splash');
 const initialNavState = RootNavigator.router.getStateForAction(splashAction);
 
-function navigationReducer(state = initialNavState, action) {
+function NavigationReducer(state = initialNavState, action) {
     let nextState = RootNavigator.router.getStateForAction(action, state);
     return nextState || state;
 }
 
-export default navigationReducer;
+export default NavigationReducer;
